scripts/ingest/requirements.txt
rasterio
geopandas
psycopg2-binary
python-dotenv
shapely
